+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # GPG Sync Changelog
 
+## 0.3.4
+
+* Uses modern keyserver, keys.openpgp.org, by default
+* Upgrades depdendencies
+* Renames --debug to --verbose, -v
+* Improvements to Windows installer
+
 ## 0.3.3
 
 * Improve macOS packaging (again)